rename Select.module style file
