+++ components/CustomerEditForm.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, SetStateAction, useState } from 'react'
 import { Customer } from '../pages/plus'
-import styles from './CustomerSelect.module.scss'
+import styles from './Select.module.scss'
 import stylesH from '../styles/Home.module.css'
 import fetchHandler, { FetchArgs } from './fetchHandler'
 type editFormArgs = {
