+++ components/EitemSelect.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, Dispatch, SetStateAction, useRef, useState } from 'react'
 import type { Eitem } from '../pages/minus'
-import styles from './CustomerSelect.module.scss'
+import styles from './Select.module.scss'
 import stylesH from '../styles/Home.module.css'
 
 type EitemSelectProps = {
