+++ components/ProductEditForm.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import { Product } from '../pages/plus'
-import styles from './CustomerSelect.module.scss'
+import styles from './Select.module.scss'
 import stylesH from '../styles/Home.module.css'
 import fetchHandler, { FetchArgs } from './fetchHandler'
 type editFormArgs = {
