+++ components/ProductSelect.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, Dispatch, SetStateAction, useRef, useState } from 'react'
 import type { Product } from '../pages/plus'
-import styles from './CustomerSelect.module.scss'
+import styles from './Select.module.scss'
 import stylesH from '../styles/Home.module.css'
 
 type ProductSelectProps = {
